feat: 기본 결제 기능 구현

+++ src/main/java/com/ll/ticket/domain/concert/entity/Concert.java
@@ -36,7 +36,7 @@ public class Concert extends BaseEntity {
 
     @Enumerated(EnumType.STRING)
     private ConcertStatus status;
-    private int seatPrice;
+    private Long seatPrice;
 
     public void setStatus(ConcertStatus status) {
         this.status = status;
